backend/requirements.txt
pandas
fastapi
uvicorn
sentence-transformers
mysql-connector-python
streamlit
pydantic
python-dotenv
psycopg2-binary>=2.9.9
tavily-python